cafe-booking/requirements.txt
asgiref==3.8.1
boto3==1.38.24
botocore==1.38.24
Django==5.2.1
django-storages==1.14.6
jmespath==1.0.1
mysqlclient==2.2.7
pillow==11.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.4.0
gunicorn==23.0.0
